refactor(selector-controller): rename select-controller.ts to selector-controller.ts
