+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from './select-controller.js';
+export * from './selector-controller.js';
 export * from './select-state.js';
